Separate cache construction & selection out of type_cache.rs into submodules
